Move Base64 section
